Refactor validation_operation_results
This adds the ability to override the key and values
that are checked against.

This is needed in particular for rename.rename_and_append
as there is a check on the addAlias which does not follow
the "status":"ok" pattern that other operation results follow.

And unfortunately there appears to be a bug in the archiver that
the addAlias can fail and the overall status is ok :(