Verify old and new pvs do not have the same data
Inparticular, before doing rename and append.

Seems like if you try do a rename and append and the data is the same
the operation fails. In this particular case it makes more sense to
delete the old pv and just add the alias, without the moving of data
in the normal rename_and_append step.